LayoutTests/imported/w3c/web-platform-tests/html/dom/elements/global-attributes/custom-attrs-expected.txt

Element Custom Attributes


FAIL Setting an Element's dataset property should not interfere with namespaced attributes with same name assert_equals: expected 3 but got 2
